Moved AffectedNodes from dynamics/ to distance/
