+++ networkit/cpp/dynamics/CMakeLists.txt
@@ -1,5 +1,4 @@
 networkit_add_module(dynamics
-    AffectedNodes.cpp
     DGSStreamParser.cpp
     DGSWriter.cpp
     GraphDifference.cpp
@@ -13,7 +12,7 @@ networkit_module_link_libraries(dynamics
     PRIVATE OpenMP::OpenMP_CXX)
 
 networkit_module_link_modules(dynamics
-        auxiliary base distance graph)
+        auxiliary base graph)
 
 add_subdirectory(test)
